requirements.txt
requirements.txt
streamlit
requests
python-dotenv
openai
